Refactors header out into menu and new-questions molecule components
